+++ src/components/banner/Info/index.tsx
