+++ src/components/banner/List/index.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import BannerItem from './BannerItem';
-import styles from '../../styles/components/banner/BannerList.module.scss';
+import BannerItem from '../Item';
 import { BannerItemType } from '@/types/banner';
-
+import styles from './styles.module.scss';
 function BannerList({ bannerList }: { bannerList: BannerItemType[] }) {
   return (
     <div className={styles.container}>
