+++ src/components/banner/List/styles.module.scss
