+++ src/components/banner/Paginate/styles.module.scss
