+++ src/components/banner/Slider/index.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
-import styles from '../../styles/components/banner/BannerSlider.module.scss';
+import styles from './styles.module.scss';
 function BannerSlider() {
   const settings = {
     dots: false,
