+++ src/components/banner/Slider/styles.module.scss
