+++ src/pages/banner/detail/styles.module.scss
