+++ src/pages/banner/list/styles.module.scss
